incorporate hud into player namespace
rearranged files, incorporated hud into player
